Process 'alg' folder
