+++ lib/index.ts
@@ -38,7 +38,9 @@ export * from './geometry/relations3';
 export * from './geometry/spherical.types';
 export * from './geometry/spherical';
 export * from './alg/binary-heap';
+export * from './alg/kd-tree.types';
 export * from './alg/kd-tree';
+export * from './alg/figures.types';
 export * from './alg/figures';
 export * from './gl/color.types';
 export * from './gl/color';
